Create find-rotation-value.go
Given a rotated but otherwise sorted array, find a particular value. This is done using two binary searches, first to find the rotation point, next to find the value in a virtual array that is constructed by mapping the indices. This is a great example of the flexibility of sort.Search().