Create shortest-reach-graph.c
https://www.hackerrank.com/challenges/ctci-bfs-shortest-reach/problem
BFS (breadth-first-search0

Consider an undirected graph consisting of n nodes where each node is labeled from 1 to n and the edge between any two nodes is always of length 6. We define node s to be the starting position for a BFS.

Given q queries in the form of a graph and some starting node s, perform each query by calculating the shortest distance from starting node s to all the other nodes in the graph. Then print a single line of n-1 space-separated integers listing node s's shortest distance to each of the n-1 other nodes (ordered sequentially by node number); if  is disconnected from a node, print -1 as the distance to that node.