Create staircase.c
https://www.hackerrank.com/challenges/ctci-recursive-staircase/problem

Davis has s staircases in his house and he likes to climb each staircase 1, 2, or 3 steps at a time. Being a very precocious child, he wonders how many ways there are to reach the top of the staircase.

Given the respective heights for each of the s staircases in his house, find and print the number of ways he can climb each staircase on a new line.